+++ src/Exception/HttpClientException.php
@@ -16,10 +16,6 @@ namespace Origin\HttpClient\Exception;
 
 use Exception;
 
-/**
- * This exception is for files that are not found when
- * using them. (This is not for Http NotFoundException)
- */
-class NotFoundException extends Exception
+class HttpClientException extends Exception
 {
 }