+++ src/Exception/ServerErrorException.php
@@ -14,12 +14,9 @@
 declare(strict_types = 1);
 namespace Origin\HttpClient\Exception;
 
-use Exception;
-
 /**
- * This exception is for files that are not found when
- * using them. (This is not for Http NotFoundException)
+ * This captures bad requests caused by server problems error 5xx
  */
-class NotFoundException extends Exception
+class ServerErrorException extends HttpException
 {
 }