+++ src/Exception/TooManyRedirectsException.php
@@ -14,12 +14,6 @@
 declare(strict_types = 1);
 namespace Origin\HttpClient\Exception;
 
-use Exception;
-
-/**
- * This exception is for files that are not found when
- * using them. (This is not for Http NotFoundException)
- */
-class NotFoundException extends Exception
+class TooManyRedirectsException extends RequestException
 {
 }