+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.0.0] - 2020-06-05
+
+### Added
+
+- Added `httpErrors` option to constructor
+- Added `HttpClientException`
+- Added `RequestException` which extends `HttpClientException`
+- Added `ConnectionException` which extends `RequestException`
+- Added `TooManyRedirectsException` which extends `RequestException`
+- Added `HttpException` which extends `RequestException`
+- Added `ClientErrorException` which extends `HttpException`
+- Added `ServerErrorException` which extends `HttpException`
+
+### Changed
+
+- Changed HTTP protocol errors trigger either Client (4xx) or Server (5xx) exception by default. This is a breaking change, 
+and can be disabled by setting the `httpErrors` when creating the `Http` object.
+
 ## [1.0.3] - 2020-05-24
 
 ### Fixed